+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Features
 
   * Auto-install CLI plugins for running services during `dcos cluster setup` when environment variable `DCOS_CLI_EXPERIMENTAL_AUTOINSTALL_PACKAGE_CLIS` set
+  * New command `dcos config keys` printing all the keys that can be set in a configuration file